src/server/workers/requirements.txt
celery
redis
python-dotenv
motor
pymongo
qwen-agent
httpx
python-jose[cryptography]
cryptography
python-dateutil
google-api-python-client
google-auth-oauthlib
google-genai